pallets are move by orientation = acc + magnetic field

+++ src/main/java/models/player/Coordinates.java
@@ -2,27 +2,27 @@ package models.player;
 
 public class Coordinates {
 
-	private int x;
-	private int y;
+	private float x;
+	private float y;
 
-	public Coordinates(int x, int y) {
+	public Coordinates(float x, float y) {
 		this.x = x;
 		this.y = y;
 	}
 
-	public int getX() {
+	public float getX() {
 		return x;
 	}
 
-	public void setX(int x) {
+	public void setX(float x) {
 		this.x = x;
 	}
 
-	public int getY() {
+	public float getY() {
 		return y;
 	}
 
-	public void setY(int y) {
+	public void setY(float y) {
 		this.y = y;
 	}
